# replit-project/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ContactMessageInputReason = typeof ContactMessageInputReason[keyof typeof ContactMessageInputReason];


export const ContactMessageInputReason = {
  Comment: 'Comment',
  Question: 'Question',
  Partnership: 'Partnership',
  Opportunity: 'Opportunity',
  Other: 'Other',
} as const;

export interface ContactMessageInput {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  email: string;
  reason: ContactMessageInputReason;
  /** @minLength 1 */
  message: string;
}

export type ContactMessageReason = typeof ContactMessageReason[keyof typeof ContactMessageReason];


export const ContactMessageReason = {
  Comment: 'Comment',
  Question: 'Question',
  Partnership: 'Partnership',
  Opportunity: 'Opportunity',
  Other: 'Other',
} as const;

export interface ContactMessage {
  id: string;
  firstName: string;
  lastName: string;
  email: string;
  reason: ContactMessageReason;
  message: string;
  submittedAt: string;
  replied: boolean;
  /** @nullable */
  repliedAt: string | null;
}

export interface ErrorEnvelope {
  error: string;
}