# replit-project/lib/api-zod/src/generated/types/contactMessage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContactMessageReason } from './contactMessageReason';

export interface ContactMessage {
  id: string;
  firstName: string;
  lastName: string;
  email: string;
  reason: ContactMessageReason;
  message: string;
  submittedAt: Date;
  replied: boolean;
  /** @nullable */
  repliedAt: Date | null;
}