# replit-project/lib/api-zod/src/generated/types/contactMessageInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContactMessageInputReason } from './contactMessageInputReason';

export interface ContactMessageInput {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  email: string;
  reason: ContactMessageInputReason;
  /** @minLength 1 */
  message: string;
}